requirements.txt
streamlit
PyPDF2
pyyaml
